almost done one solution to test only

+++ GarageProject/GarageProject/Vehicles/Airplane.cs
@@ -9,10 +9,10 @@ namespace GarageProject.Vehicles
 {
     public class Airplane : Vehicle
     {
-        private int wingsLength;
+        private float wingsLength;
         private int numberOfWindoors;
 
-        public int WingsLength { set { wingsLength = value; } get { return wingsLength; } }
+        public float WingsLength { set { wingsLength = value; } get { return wingsLength; } }
         public int NumberOfWindoors { set { numberOfWindoors = value; } get { return numberOfWindoors; } }
 
         public Airplane()
@@ -20,10 +20,11 @@ namespace GarageProject.Vehicles
 
         }
 
-        public Airplane(int id, string fName, string lName, string registrationNumber, string type, string model, int numberOfWheels, int numberOfSeats, int numberOfEngines,
-            string color, DateTime parkingTime, DateTime unparkingTime, float parkingPeriod, int wingsLength, int numberOfWindoors):
-            base(id, fName, lName, registrationNumber, type, model,
-           color, numberOfWheels, numberOfSeats, numberOfEngines, parkingTime, unparkingTime, parkingPeriod)
+        public Airplane(int id, string fName, string lName, string registrationNumber, string type, string model, 
+                        int numberOfWheels, int numberOfSeats, int numberOfEngines, string color, DateTime parkingTime, 
+                        DateTime unparkingTime, float parkingPeriod, float wingsLength, int numberOfWindoors) :
+        base(id, fName, lName, registrationNumber, type, model, color, numberOfWheels,
+             numberOfSeats, numberOfEngines, parkingTime, unparkingTime, parkingPeriod)
         {
             WingsLength = wingsLength;
             NumberOfWindoors = numberOfWindoors;
@@ -33,6 +34,5 @@ namespace GarageProject.Vehicles
         {            
         return base.VehicleInformations() + "\nWings Length is: " + WingsLength + "\nNumber of Windoors is: " + NumberOfWindoors;
         }
-
     }
 }
